add custom dataloader that allows for more flexible training

+++ NILM_Dataset.py
@@ -22,9 +22,9 @@ class NILMDataset(data_utils.Dataset):
         y        = self.padding_seqs(self.y[start_index: end_index])
         status   = self.padding_seqs(self.status[start_index: end_index])
         
-        x        = torch.tensor(x).view((1,-1))
-        y        = torch.tensor(y).view((1,-1))
-        status   = torch.tensor(status).view((1,-1))
+        x        = torch.tensor(x).view((1,-1)).float()
+        y        = torch.tensor(y).view((1,-1)).float()
+        status   = torch.tensor(status).view((1,-1)).float()
         
         return x, y, status
